LayoutTests/fast/eventloop/queue-task-event-loop-starvation-expected.txt

This tests repeated queueing of event loop tasks. Event loop should eventually yield to the run loop.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS executedTaskCount < 100 is true
PASS successfullyParsed is true

TEST COMPLETE
